Refactor : 핀 만들기 - pin-make.js 로 파일 분리해놨던 것 pin.js로 합침
