+++ routes/pins.js
@@ -1,7 +1,5 @@
 import express from 'express';
 import * as pinsController from '../controllers/pins.js';
-import * as pinsMakeController from '../controllers/pin-make.js';
-import * as pinsOrganizeController from '../controllers/pin-organize.js';
 import { isLogin } from '../middleware/auth.js';
 import multer from 'multer';
 import path from 'path';
@@ -33,8 +31,8 @@ router.use(isLogin);
 router.get('/pins', pinsController.pinList);
 router.get('/pins/:pin_id', pinsController.readPinById);
 router.post('/pins/:pin_id', pinsController.savePin);
-router.get('/pin-make', pinsMakeController.readMakePinPage);
-router.post('/pin-make', upload.single('image'), pinsMakeController.createPin);
+router.get('/pin-make', pinsController.readMakePinPage);
+router.post('/pin-make', upload.single('image'), pinsController.createPin);
 router.post('/pin-organize', pinsController.organizePins);
 
 export default router;
